+++ CalendarScheduleGenerator2/Groupe.cs
@@ -1,0 +1,15 @@
+// list classe + site
+// --------------------------------------------
+// --------------------------------------------
+// --------------------------------------------
+
+// groupe
+namespace CalendarScheduleGenerator2
+{
+    public class Groupe
+    {
+        public string Name { get; set; } = string.Empty;
+        public int Capacity { get; set; } = 0;
+        public string PreferedSite { get; set; } = string.Empty;
+    }
+}
